[refactor] 전시회 검색창 재사용 틀 분리
