+++ src/screens/mydiary/MyExhSearchScreen.tsx
@@ -1,23 +1,14 @@
-import React, {useCallback, useState} from 'react';
-import {Keyboard, TouchableOpacity} from 'react-native';
+import React, {useState} from 'react';
+import {Keyboard} from 'react-native';
 import styled from 'styled-components/native';
-import {SearchIcon} from '~/assets/images';
 import BackView from '~/components/common/BackView';
-import {
-  widthPercentage as wp,
-  heightPercentage as hp,
-  fontPercentage as fp,
-} from '~/components/common/ResponsiveSize';
-import SearchExhList from '../../components/exhSearch/SearchExhList';
+import SearchExhList from './SearchExhList';
 import {showToast} from '~/components/common/modal/toastConfig';
+import SearchExhFrame from '~/components/exhSearch/SearchExhFrame';
 
 const MyExhSearchScreen = () => {
   const [searchKeyword, setSearchKeyword] = useState<string>('');
   const [keyword, setKeyword] = useState<string>('');
-
-  const onChangeText = useCallback((text: string) => {
-    setSearchKeyword(text);
-  }, []);
 
   const onPressSearch = () => {
     if (checkKeyword(searchKeyword)) {
@@ -42,22 +33,11 @@ const MyExhSearchScreen = () => {
   return (
     <Container>
       <BackView line={false} children={null} />
-      <ContentsContainer>
-        {/* 검색창 */}
-        <SearchContainer>
-          <SearchView>
-            <SearchWord
-              onSubmitEditing={onPressSearch}
-              onChangeText={onChangeText}
-              placeholder={'전시회를 검색하세요'}
-              value={searchKeyword}
-            />
-            <TouchableOpacity onPress={onPressSearch}>
-              <SearchIcon />
-            </TouchableOpacity>
-          </SearchView>
-          <UnderLine />
-        </SearchContainer>
+      {/* 검색창 */}
+      <SearchExhFrame
+        searchKeyword={searchKeyword}
+        onPressSearch={onPressSearch}
+        handleSearchKeyword={setSearchKeyword}>
         {/* 전시회 목록 */}
         {keyword !== '' && (
           <SearchExhList
@@ -65,7 +45,7 @@ const MyExhSearchScreen = () => {
             changeIsPressed={onPressSearch}
           />
         )}
-      </ContentsContainer>
+      </SearchExhFrame>
     </Container>
   );
 };
@@ -78,40 +58,4 @@ const Container = styled.View`
   flex-direction: column;
   width: 100%;
   background-color: #f6f6f6;
-`;
-
-const ContentsContainer = styled.View`
-  flex: 1;
-  flex-direction: column;
-  width: 100%;
-  padding-top: ${hp(5)}px;
-  padding-bottom: ${hp(5)}px;
-  padding-left: ${wp(15)}px;
-  padding-right: ${wp(15)}px;
-`;
-
-const SearchContainer = styled.View`
-  flex-direction: column;
-  padding-top: ${hp(5)}px;
-  padding-bottom: ${hp(5)}px;
-  padding-left: ${wp(10)}px;
-  padding-right: ${wp(10)}px;
-`;
-
-const UnderLine = styled.View`
-  border-bottom-width: 1px;
-  border-color: #ff6f61;
-`;
-
-const SearchView = styled.View`
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const SearchWord = styled.TextInput`
-  flex: 1;
-  font-size: ${fp(16)}px;
-  color: #3c4045;
-  font-family: 'omyu pretty';
 `;
